feat(backend): basic tests implemeneted for login, register and logout
A series of tests have been created using vitest and supertest to verify the functionalities
implemented in the backend. Specifically, the login, register, and logout operations have been
tested. ISSUE: It is not possible to perform tests to verify the correct functionality; tests can
only be conducted to identify errors.